requirements.txt
numpy
pandas
scikit-learn
PyWavelets
joblib
requests
PyYAML
pytest
responses
